My first ever Travis CI build using databases.
Going to attempt to build on Travis CI using the PostgreSQL and MySQL database
resources available on Travis CI. Added a `.configuration.travis.json` to the
project that has configurations for the MySQL and PostgreSQL at Travis CI. It
gets moved into place as part of the `before_install` in `.travis.yml`.

The `.travis.yml` will build a `relatable` database in MySQL and PostgreSQL
as part of `before_install`.