My first ever Travis CI build using databases.
Building Relatable at Travis CI, testing against the Travis CI provided
MySQL and PostgresSQL.

Added a `.configuration.travis.json` to the project that has
configurations for the MySQL and PostgreSQL at Travis CI. It gets moved
into place as part of the `before_install` in `.travis.yml`.

The `.travis.yml` will build a `relatable` database in MySQL and PostgreSQL
as part of `before_install`. The `before_install` setup  uses the DDL in
the `ddl` directory to create the test tables.

The Twinkie submodule was the read/write URL, requiring a private key
for authentication that Travis CI does not have. The submodule now uses
the read only `git` URL.

Renamed `ddl/mysql.ddl` to `ddl/mysql.sql` to be consistant with
`ddl/postgresql.sql`.