Implement sub-select in SELECT clause minimally.
Now able to specify a sub-select in a SELECT clause. This is a minimal
implementation. It is not any less powerful than the current
implementation, where sub-selects are subsequent queries, but it is no
more powerful either. Notably, I do not believe that sub-selects will
nest beyond a single level.

Closes #32.