Scan parameters, sub-selects only.
Compile parameters into the query structure. I'm gathering parameters
into an array of functions. The functions return a value out of a
context object. I'm inserted `?` placeholders for each parameters, but I
need to parameterize the query compilation function to accept a strategy
for generating placeholders because the syntax is platform specific.

In this commit, I got rid of trees defined by a subsequent query.